Moved analysis tests into trafo tests file, to reduce test overhead.
